Restructure html file, fix css, add image
The html structure needs some modifications to the outer and inner containers for better layout. On the css styling, I used flexbox for an easier layout of the components. For the script.js modications, a lot of appending was done to be able see the desired outcome, this may not be the best practice but for now I will settle to it so that I can proceed on to fixing the main logic of the game.